+++ CHANGELOG.md
@@ -30,6 +30,12 @@ and this project adheres to
 
 ### Changed
 
+- AI Assistant: Show disclaimer once every day per user
+  [#2481](https://github.com/OpenFn/lightning/issues/2481)
+- AI Assistant: Scroll to new message when it arrives
+  [#2409](https://github.com/OpenFn/lightning/issues/2409)
+- AI Assistant: Set vertical scrollbar below the session title
+  [#2477](https://github.com/OpenFn/lightning/issues/2477)
 - Bumped dependencies
 - Extend display of audit events to cater for deletions.
   [#2701](https://github.com/OpenFn/lightning/issues/2701)
@@ -50,7 +56,7 @@ and this project adheres to
 - Allow filtering collection items by updated_before and updated_after.
   [#2693](https://github.com/OpenFn/lightning/issues/2693)
 - Add support for SMTP email configuration
-  [#2699](https://github.com/OpenFn/lightning/issues/2699)  
+  [#2699](https://github.com/OpenFn/lightning/issues/2699)
   ⚠️️ Please note that `EMAIL_ADMIN` defaults to `lightning@example.com` in
   production environments